QR code Generator/requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pillow==11.0.0
pyzbar==0.1.9
qrcode==8.0
Werkzeug==3.1.3
